+++ openbas-front/src/admin/components/atomic_testings/atomic_testing/IndexTitle.tsx
@@ -1,11 +1,10 @@
 import { InfoOutlined } from '@mui/icons-material';
 import { Box, IconButton, Tooltip, Typography } from '@mui/material';
-import { useContext } from 'react';
 import { makeStyles } from 'tss-react/mui';
 
 import Loader from '../../../../components/Loader';
+import type { InjectResultOverviewOutput } from '../../../../utils/api-types';
 import { truncate } from '../../../../utils/String';
-import { InjectResultOverviewOutputContext, type InjectResultOverviewOutputContextType } from '../InjectResultOverviewOutputContext';
 import AtomicTestingInformation from './AtomicTestingInformation';
 
 const useStyles = makeStyles()(() => ({
@@ -19,25 +18,25 @@ const useStyles = makeStyles()(() => ({
   },
 }));
 
-const AtomicTestingHeader = () => {
+interface Props { injectResultOverview: InjectResultOverviewOutput }
+
+const IndexTitle = ({ injectResultOverview }: Props) => {
   // Standard hooks
   const { classes } = useStyles();
 
-  const { injectResultOverviewOutput } = useContext<InjectResultOverviewOutputContextType>(InjectResultOverviewOutputContext);
-
-  if (!injectResultOverviewOutput) {
+  if (!injectResultOverview) {
     return <Loader variant="inElement" />;
   }
 
   return (
     <Box>
-      <Tooltip title={injectResultOverviewOutput.inject_title}>
+      <Tooltip title={injectResultOverview.inject_title}>
         <Typography
           variant="h1"
           gutterBottom
           classes={{ root: classes.title }}
         >
-          {truncate(injectResultOverviewOutput.inject_title, 80)}
+          {truncate(injectResultOverview.inject_title, 80)}
         </Typography>
       </Tooltip>
       <Tooltip
@@ -47,12 +46,12 @@ const AtomicTestingHeader = () => {
               maxWidth: 500,
               backgroundColor: 'background.paper',
               color: 'text.primary',
-              padding:0,
+              padding: 0,
             },
           },
         }}
         title={
-          <AtomicTestingInformation injectResultOverviewOutput={injectResultOverviewOutput} />
+          <AtomicTestingInformation injectResultOverviewOutput={injectResultOverview} />
         }
       >
         <IconButton size="small">
@@ -66,4 +65,4 @@ const AtomicTestingHeader = () => {
   );
 };
 
-export default AtomicTestingHeader;
+export default IndexTitle;
