+++ openbas-front/src/admin/components/simulations/simulation/injects/InjectIndex.tsx
@@ -1,40 +1,19 @@
-import { Box, Tab, Tabs } from '@mui/material';
 import { type FunctionComponent, Suspense, useEffect, useState } from 'react';
-import { Link, Route, Routes, useLocation, useParams, useSearchParams } from 'react-router';
-import { makeStyles } from 'tss-react/mui';
+import { useParams } from 'react-router';
 
 import { fetchInjectResultOverviewOutput } from '../../../../../actions/atomic_testings/atomic-testing-actions';
 import { fetchExercise } from '../../../../../actions/Exercise';
 import { type ExercisesHelper } from '../../../../../actions/exercises/exercise-helper';
-import Breadcrumbs, { BACK_LABEL, BACK_URI, type BreadcrumbsElement } from '../../../../../components/Breadcrumbs';
-import { errorWrapper } from '../../../../../components/Error';
-import { useFormatter } from '../../../../../components/i18n';
 import Loader from '../../../../../components/Loader';
-import NotFound from '../../../../../components/NotFound';
 import { useHelper } from '../../../../../store';
 import { type Exercise as ExerciseType, type InjectResultOverviewOutput } from '../../../../../utils/api-types';
 import { usePermissions } from '../../../../../utils/Exercise';
 import { useAppDispatch } from '../../../../../utils/hooks';
 import useDataLoader from '../../../../../utils/hooks/useDataLoader';
-import AtomicTesting from '../../../atomic_testings/atomic_testing/AtomicTesting';
-import AtomicTestingDetail from '../../../atomic_testings/atomic_testing/AtomicTestingDetail';
-import AtomicTestingFindings from '../../../atomic_testings/atomic_testing/AtomicTestingFindings';
-import AtomicTestingPayloadInfo from '../../../atomic_testings/atomic_testing/payload_info/AtomicTestingPayloadInfo';
+import IndexRoutes from '../../../atomic_testings/atomic_testing/IndexRoutes';
 import { InjectResultOverviewOutputContext } from '../../../atomic_testings/InjectResultOverviewOutputContext';
 import { PermissionsContext, type PermissionsContextType } from '../../../common/Context';
-import InjectHeader from '../../../injects/InjectHeader';
-
-const useStyles = makeStyles()(() => ({
-  item: {
-    height: 30,
-    fontSize: 13,
-    float: 'left',
-    whiteSpace: 'nowrap',
-    overflow: 'hidden',
-    textOverflow: 'ellipsis',
-    paddingRight: 10,
-  },
-}));
+import InjectIndexHeader from './InjectIndexHeader';
 
 const InjectIndexComponent: FunctionComponent<{
   exercise: ExerciseType;
@@ -43,52 +22,13 @@ const InjectIndexComponent: FunctionComponent<{
   exercise,
   injectResult,
 }) => {
-  // Standard hooks
-  const { t } = useFormatter();
-  const { classes } = useStyles();
-
   // Context
   const permissionsContext: PermissionsContextType = { permissions: usePermissions(exercise.exercise_id) };
-
-  const [searchParams] = useSearchParams();
-  const backlabel = searchParams.get('backlabel');
-  const backuri = searchParams.get('backuri');
-  const location = useLocation();
-  const tabValue = location.pathname;
 
   const [injectResultOverviewOutput, setInjectResultOverviewOutput] = useState<InjectResultOverviewOutput>(injectResult);
 
   const updateInjectResultOverviewOutput = (newData: InjectResultOverviewOutput) => {
     setInjectResultOverviewOutput(newData);
-  };
-
-  const breadcrumbs: BreadcrumbsElement[] = [
-    {
-      label: t('Simulations'),
-      link: '/admin/simulations',
-    },
-    {
-      label: t(exercise.exercise_name),
-      link: `/admin/simulations/${exercise.exercise_id}`,
-    },
-  ];
-  if (backlabel && backuri) {
-    breadcrumbs.push({
-      label: backlabel,
-      link: backuri,
-    });
-  }
-  breadcrumbs.push({ label: t('Injects') });
-  breadcrumbs.push({
-    label: injectResultOverviewOutput.inject_title,
-    current: true,
-  });
-
-  const computePath = (path: string) => {
-    if (backlabel && backuri) {
-      return path + `?${BACK_LABEL}=${backlabel}&${BACK_URI}=${backuri}`;
-    }
-    return path;
   };
 
   return (
@@ -98,74 +38,9 @@ const InjectIndexComponent: FunctionComponent<{
     }}
     >
       <PermissionsContext.Provider value={permissionsContext}>
-        <div style={{
-          display: 'flex',
-          flexDirection: 'row',
-          gap: '20px',
-        }}
-        >
-          <div style={{ flex: 1 }}>
-            <Breadcrumbs variant="object" elements={breadcrumbs} />
-            <InjectHeader inject={injectResult} />
-          </div>
-        </div>
-        <Box
-          sx={{
-            borderBottom: 1,
-            borderColor: 'divider',
-            marginBottom: 4,
-          }}
-        >
-          <Tabs value={tabValue}>
-            <Tab
-              component={Link}
-              to={computePath(`/admin/simulations/${exercise.exercise_id}/injects/${injectResultOverviewOutput.inject_id}`)}
-              value={`/admin/simulations/${exercise.exercise_id}/injects/${injectResultOverviewOutput.inject_id}`}
-              label={t('Overview')}
-              className={classes.item}
-            />
-            {(injectResultOverviewOutput.inject_injector_contract?.injector_contract_payload
-              || injectResultOverviewOutput.inject_type === 'openbas_nmap') && (
-              <Tab
-                component={Link}
-                to={computePath(`/admin/simulations/${exercise.exercise_id}/injects/${injectResultOverviewOutput.inject_id}/findings`)}
-                value={`/admin/simulations/${exercise.exercise_id}/injects/${injectResultOverviewOutput.inject_id}/findings`}
-                label={t('Findings')}
-                className={classes.item}
-              />
-            )}
-            <Tab
-              component={Link}
-              to={computePath(`/admin/simulations/${exercise.exercise_id}/injects/${injectResultOverviewOutput.inject_id}/detail`)}
-              value={`/admin/simulations/${exercise.exercise_id}/injects/${injectResultOverviewOutput.inject_id}/detail`}
-              label={t('Execution details')}
-              className={classes.item}
-            />
-            {
-              injectResultOverviewOutput.inject_injector_contract?.injector_contract_payload && (
-                <Tab
-                  component={Link}
-                  to={computePath(`/admin/simulations/${exercise.exercise_id}/injects/${injectResultOverviewOutput.inject_id}/payload_info`)}
-                  value={`/admin/simulations/${exercise.exercise_id}/injects/${injectResultOverviewOutput.inject_id}/payload_info`}
-                  label={t('Payload info')}
-                  className={classes.item}
-                />
-              )
-            }
-          </Tabs>
-        </Box>
+        <InjectIndexHeader injectResultOverview={injectResultOverviewOutput} exercise={exercise} />
         <Suspense fallback={<Loader />}>
-          <Routes>
-            <Route path="" element={errorWrapper(AtomicTesting)()} />
-            { (injectResultOverviewOutput.inject_injector_contract?.injector_contract_payload
-              || injectResultOverviewOutput.inject_type === 'openbas_nmap')
-            && <Route path="findings" element={errorWrapper(AtomicTestingFindings)()} />}
-            <Route path="detail" element={errorWrapper(AtomicTestingDetail)()} />
-            { injectResultOverviewOutput.inject_injector_contract?.injector_contract_payload
-              && <Route path="payload_info" element={errorWrapper(AtomicTestingPayloadInfo)()} />}
-            {/* Not found */}
-            <Route path="*" element={<NotFound />} />
-          </Routes>
+          <IndexRoutes injectResultOverview={injectResultOverviewOutput} />
         </Suspense>
       </PermissionsContext.Provider>
     </InjectResultOverviewOutputContext.Provider>
